# openapi/payments.yaml
openapi: 3.0.3
info:
  title: Payments API
  version: 1.0.0
  description: |
    Demo payments service. Sensitive mutations require identity verification.
servers:
  - url: https://apim.example.com
paths:
  /payments/direct-debit:
    post:
      summary: Update direct debit details
      operationId: updateDirectDebit
      tags: [payments]
      parameters:
        - in: header
          name: x-correlation-id
          required: true
          schema: { type: string }
        - in: header
          name: x-verification-token
          required: true
          schema: { type: string }
          description: Issued by IDV service; enforced by APIM
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DirectDebitUpdateRequest'
            examples:
              happy:
                value:
                  customer_id: "cust-001"
                  iban: "BE68539007547034"
                  mandate_reference: "MDT-000123"
      responses:
        '200':
          description: Updated
          headers:
            x-correlation-id:
              required: true
              schema: { type: string }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DirectDebitUpdateResponse'
              examples:
                ok:
                  value:
                    status: "success"
                    updated_at: "2026-02-17T12:00:00Z"
        '400':
          description: Invalid request
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '401':
          description: Unauthorized
        '403':
          description: Forbidden (verification required or not authorized)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              examples:
                verification_required:
                  value:
                    code: "verification_required"
                    message: "Identity verification token is required."
  /payments/direct-debit/{customer_id}:
    get:
      summary: Get direct debit details for a customer
      operationId: getDirectDebit
      tags: [payments]
      parameters:
        - in: header
          name: x-correlation-id
          required: true
          schema: { type: string }
        - in: path
          name: customer_id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Direct debit record
          headers:
            x-correlation-id:
              required: true
              schema: { type: string }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DirectDebitRecord'
        '404':
          description: Not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

components:
  schemas:
    DirectDebitUpdateRequest:
      type: object
      required: [customer_id, iban, mandate_reference]
      additionalProperties: false
      properties:
        customer_id: { type: string }
        iban:
          type: string
          description: IBAN (demo format validation handled by service/APIM if desired)
        mandate_reference: { type: string }
    DirectDebitUpdateResponse:
      type: object
      required: [status, updated_at]
      additionalProperties: false
      properties:
        status:
          type: string
          enum: [success]
        updated_at:
          type: string
          format: date-time
    DirectDebitRecord:
      type: object
      required: [customer_id, iban, mandate_reference, updated_at]
      additionalProperties: false
      properties:
        customer_id: { type: string }
        iban: { type: string }
        mandate_reference: { type: string }
        updated_at: { type: string, format: date-time }    
    Error:
      type: object
      required: [code, message]
      additionalProperties: false
      properties:
        code: { type: string }
        message: { type: string }